requirements.txt
fastapi[all]>=0.111.0
uvicorn>=0.30.0
sqlalchemy[asyncio]>=2.0.31
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.1
pytest>=8.2.2
httpx>=0.27.0
